enable to recursively copy directories to remote
Machinectl cannot overwrite existing files with `copy-to`.  There is a
pending RFE (systemd/systemd#9441), but currently no other alternative
(to the best of my knowledge) than proactively remove the existing file
before pushing a new one.

This is especially annoying when copying a whole directory. As this is
done recursively file by file, and since the ansible script pushed on
the remote has the same path, this currently triggers machinectl which
rejects the copy with a "file exists" message.

The kludge proposed in this path is to systematically attempt to
remove existing files before pushing them to the remote... That's
definitely not great, but I have not found better...